Item classes, see #147 Add GameMode enumerator Add license headers Move DamageItem() from Inventory to the item class

+++ Chraft/Net/Client.cs
@@ -155,7 +155,7 @@ namespace Chraft.Net
         {
             SendPacket(new NewInvalidStatePacket
             {
-                GameMode = _player.GameMode,
+                GameMode = (byte)_player.GameMode,
                 Reason = NewInvalidStatePacket.NewInvalidReason.ChangeGameMode
             });
         }
